backend/requirements.txt
alembic==1.13.0
boto3==1.34.0
jinja2==3.1.2
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
cffi==2.0.0
click==8.1.7
cryptography==45.0.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
